refactor: Mbti 페이지 컴포넌트 통합
